+++ docs/changelog.md
@@ -26,6 +26,7 @@ Ends August 31, 2020
 - add END_TURN action.
 - add new Card to Hand when Card is drawn from Deck.
 - animate End Turn ArrowIcon when turn is endable.
+- update handleIconClick in ArrowIcon component so that it's unable to return false.
 
 ## Sprint 2
 Ended August 17, 2020